Apps/Batch BG Remover/requirements.txt
customtkinter
Pillow
transparent-background
torch
numpy
tqdm
pytest
